Refactor parafac initialization
We factor out the factor initialization code from the main parafac algorithm. We
do this in anticipation of upcoming changes to the main algorithm mentioned in
PR #29. Otherwise, `parafac()` will be too large of a function and, therefore,
difficult to debug.

Added a collection of randomized parafac tests, as well.